pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paw"
version = "0.1.0"
description = "PAW — Personal Agent Workspace. A self-hosted AI agent that lives in its own Linux environment."
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "litellm>=1.55.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "structlog>=24.4.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "typer[all]>=0.15.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[project.scripts]
paw = "paw.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/paw"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
